refactor: ErrorFallback 컴포넌트를 common 폴더 하위로 이동
