+++ src/components/pages/ProductPage/ProductPage.tsx
@@ -2,7 +2,7 @@ import { Suspense } from 'react';
 import ProductList from '../../product/ProductList/ProductList';
 import ProductListFallback from '../../product/ProductList/ProductListFallback';
 import ErrorBoundary from '../../common/ErrorBoundary/ErrorBoundary';
-import ErrorFallback from '../ErrorFallback/ErrorFallback';
+import ErrorFallback from '../../common/ErrorFallback/ErrorFallback';
 
 const ProductPage = () => {
   return (
